Move the conditionals check to the tasks shared directory
