+++ tasks/realvncserver/rvstask.go
@@ -12,6 +12,7 @@ import (
 
 	tacoexec "github.com/realvnc-labs/tacoscript/exec"
 	"github.com/realvnc-labs/tacoscript/tasks"
+	"github.com/realvnc-labs/tacoscript/tasks/shared/conditionals"
 	"github.com/realvnc-labs/tacoscript/tasks/shared/executionresult"
 	"github.com/realvnc-labs/tacoscript/tasks/shared/fieldstatus"
 )
@@ -184,7 +185,7 @@ func (rvste *Executor) Execute(ctx context.Context, task tasks.CoreTask) executi
 	}
 
 	logrus.Debugf("will check if the task '%s' should be executed", task.GetPath())
-	skipReason, err := tasks.CheckConditionals(execCtx, rvste.FsManager, rvste.Runner, rvst)
+	skipReason, err := conditionals.Check(execCtx, rvste.FsManager, rvste.Runner, rvst)
 	if err != nil {
 		execRes.Err = err
 		return execRes
