+++ tasks/validate.go
@@ -5,8 +5,6 @@ import (
 	"strings"
 )
 
-const onlyIfConditionFailedReason = "only if condition was false"
-
 func ValidateRequired(val, path string) error {
 	if strings.TrimSpace(val) == "" {
 		return fmt.Errorf("empty required value at path '%s'", path)
